Change columns type from integer to double & Add discount column to orders

+++ database/migrations/2021_02_06_000001_create_programs_table.php
@@ -16,7 +16,7 @@ class CreateProgramsTable extends Migration
         Schema::create('programs', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer("hourPrice");
+            $table->double("hourPrice");
             $table->timestamps();
         });
     }
